+++ inc/toolbox.class.php
@@ -232,6 +232,12 @@ class PluginFieldsToolbox
             Socket::class,
         ];
 
+        foreach ($CFG_GLPI['asset_types'] as $asset_type) {
+            if (str_starts_with($asset_type, "Glpi\CustomAsset")) {
+                $assets_itemtypes[] = $asset_type;
+            }
+        }
+
         $assistance_itemtypes = [
             Ticket::class,
             Problem::class,